refactoring dataset store
