+++ lumigator/frontend/src/stores/datasetsStore.ts
@@ -1,36 +1,25 @@
-import { computed, ref, watch, type Ref } from 'vue'
+import { computed, ref, type Ref } from 'vue'
 import { defineStore } from 'pinia'
 import { datasetsService } from '@/sdk/datasetsService'
 
-import { useToast } from 'primevue/usetoast'
-import type { ToastMessageOptions } from 'primevue'
 import type { Dataset } from '@/types/Dataset'
-import { getAxiosError } from '@/helpers/getAxiosError'
-import type { AxiosError } from 'axios'
-import { downloadContent } from '@/helpers/downloadContent'
 import { jobsService } from '@/sdk/jobsService'
 import type { Job } from '@/types/Job'
 import { retrieveEntrypoint } from '@/helpers/retrieveEntrypoint'
-import type { EvaluationJobResults } from '@/types/Experiment'
 import { WorkflowStatus } from '@/types/Workflow'
 import { calculateDuration } from '@/helpers/calculateDuration'
 
 export const useDatasetStore = defineStore('datasets', () => {
   const datasets: Ref<Dataset[]> = ref([])
   const selectedDataset: Ref<Dataset | undefined> = ref()
-  const selectedJob: Ref<Job | undefined> = ref()
+
   const completedStatus = [WorkflowStatus.SUCCEEDED, WorkflowStatus.FAILED]
 
-  const selectedJobResults: Ref<EvaluationJobResults[]> = ref([])
   const jobs: Ref<Job[]> = ref([])
   const inferenceJobs: Ref<Job[]> = ref([])
-  const jobLogs: Ref<string[]> = ref([])
+
   const isPollingForJobStatus = ref(false)
   let jobStatusInterval: number | undefined = undefined
-  const isPollingForJobLogs = ref(false)
-  let jobLogsInterval: number | undefined = undefined
-
-  const toast = useToast()
 
   async function fetchDatasets() {
     try {
@@ -40,9 +29,9 @@ export const useDatasetStore = defineStore('datasets', () => {
     }
   }
 
-  const hasRunningInferenceJob = computed(() => {
-    return inferenceJobs.value.some((job) => job.status === WorkflowStatus.RUNNING)
-  })
+  const setSelectedDataset = (dataset: Dataset | undefined): void => {
+    selectedDataset.value = dataset
+  }
 
   /**
    * Loads all experiments and jobs.
@@ -61,6 +50,10 @@ export const useDatasetStore = defineStore('datasets', () => {
       .filter((job) => job.metadata.job_type === 'evaluate')
       .map((job) => parseJob(job))
   }
+
+  const hasRunningInferenceJob = computed(() => {
+    return inferenceJobs.value.some((job) => job.status === WorkflowStatus.RUNNING)
+  })
 
   /**
    *
@@ -82,72 +75,6 @@ export const useDatasetStore = defineStore('datasets', () => {
    */
   function getIncompleteJobIds() {
     return jobs.value.filter((job) => !completedStatus.includes(job.status)).map((job) => job.id)
-  }
-
-  async function fetchJob(id: string) {
-    const jobData = await jobsService.fetchJob(id)
-    selectedJob.value = parseJob(jobData)
-  }
-
-  async function retrieveJobLogs() {
-    if (selectedJob.value) {
-      const logsData = await jobsService.fetchLogs(selectedJob.value?.id)
-      const logs = splitByEscapeCharacter(logsData.logs)
-      logs.forEach((log: string) => {
-        const lastEntry = jobLogs.value[jobLogs.value.length - 1]
-        if (jobLogs.value.length === 0 || lastEntry !== log) {
-          jobLogs.value.push(log)
-        }
-      })
-    }
-  }
-
-  function startPollingForAnnotationJobLogs() {
-    jobLogs.value = []
-    if (!isPollingForJobLogs.value) {
-      isPollingForJobLogs.value = true
-      retrieveJobLogs()
-      // Poll every 3 seconds
-      jobLogsInterval = setInterval(retrieveJobLogs, 3000)
-    }
-  }
-
-  // start/stop polling for selected job (if its running) as the user clicks through them
-  watch(
-    selectedJob,
-    (newValue) => {
-      jobLogs.value = []
-      if (newValue) {
-        // const isEvaluationJob = jobs.value.some((job) => job?.id === newValue.id)
-        // if (isEvaluationJob) {
-        // switch to the experiment the job belongs
-        // const selectedExperimentId = experiments.value.find(experiment => {
-        //   return experiment.workflows.some(workflow => workflow.id === newValue.id)
-        // })
-        // selectedExperiment.value = experiments.value.find((exp) => exp.id === newValue.id)
-        // }
-        retrieveJobLogs()
-      }
-      if (newValue?.status === WorkflowStatus.RUNNING) {
-        startPollingForAnnotationJobLogs()
-      } else if (isPollingForJobLogs.value) {
-        stopPollingForAnnotationJobLogs()
-      }
-    },
-    { deep: true },
-  )
-
-  function stopPollingForAnnotationJobLogs() {
-    if (isPollingForJobLogs.value) {
-      isPollingForJobLogs.value = false
-      clearInterval(jobLogsInterval)
-      jobLogsInterval = undefined
-    }
-  }
-
-  function splitByEscapeCharacter(input: string) {
-    const result = input.split('\n')
-    return result
   }
 
   /**
@@ -181,7 +108,7 @@ export const useDatasetStore = defineStore('datasets', () => {
 
   function startPollingForAnnotationJobStatus(jobId: string) {
     isPollingForJobStatus.value = true
-    jobLogsInterval = setInterval(() => {
+    jobStatusInterval = setInterval(() => {
       updateJobStatus(jobId).then(() => {
         const job = jobs.value.find((job) => job.id === jobId)
         if (job && completedStatus.includes(job.status)) {
@@ -219,90 +146,16 @@ export const useDatasetStore = defineStore('datasets', () => {
     }
   }
 
-  async function fetchDatasetDetails(datasetID: string) {
-    try {
-      selectedDataset.value = await datasetsService.fetchDatasetInfo(datasetID)
-    } catch {
-      selectedDataset.value = undefined
-    }
-  }
-
-  function resetSelection() {
-    selectedDataset.value = undefined
-  }
-
-  async function uploadDataset(datasetFile: File) {
-    if (!datasetFile) {
-      return
-    }
-    // Create a new FormData object and append the selected file and the required format
-    const formData = new FormData()
-    formData.append('dataset', datasetFile) // Attach the file
-    formData.append('format', 'job') // Specification @localhost:8000/docs
-    try {
-      await datasetsService.postDataset(formData)
-    } catch (error) {
-      const errorMessage = getAxiosError(error as Error | AxiosError)
-      toast.add({
-        severity: 'error',
-        summary: `${errorMessage}`,
-        messageicon: 'pi pi-exclamation-triangle',
-        group: 'br',
-      } as ToastMessageOptions & { messageicon: string })
-    }
-    await fetchDatasets()
-  }
-
-  // async function fetchJobResults(jobId: string) {
-  //   const results = (await experimentsService.fetchExperimentResults(jobId)) as {
-  //     resultsData: WorkflowResults
-  //     id: string
-  //     download_url: string
-  //   }
-  //   if (results?.id) {
-  //     selectedJob.value = jobs.value.find((job) => job.id === results.id)
-  //     selectedJobResults.value = transformJobResults(results.resultsData)
-  //   }
-  // }
-
-  async function deleteDataset(id: string) {
-    if (!id) {
-      return
-    }
-    if (selectedDataset.value?.id === id) {
-      resetSelection()
-    }
-    await datasetsService.deleteDataset(id)
-    await fetchDatasets()
-  }
-
-  // TODO: this shouldn't depend on refs/state, it can be a util function
-  async function downloadDatasetFile() {
-    if (selectedDataset.value) {
-      const blob = await datasetsService.downloadDataset(selectedDataset.value?.id)
-      downloadContent(blob, selectedDataset.value?.filename)
-    }
-  }
-
   return {
     datasets,
-    fetchDatasets,
     selectedDataset,
-    fetchDatasetDetails,
-    jobLogs,
-    resetSelection,
-    uploadDataset,
-    deleteDataset,
-    downloadDatasetFile,
     jobs,
     inferenceJobs,
-    selectedJob,
-    selectedJobResults,
     hasRunningInferenceJob,
+    fetchDatasets,
+    setSelectedDataset,
     fetchAllJobs,
     updateStatusForIncompleteJobs,
-    fetchJob,
-    // fetchJobResults,
     startGroundTruthGeneration,
     parseJob,
   }
